Changed all the up SQL to create tables with SERIAL and INT rather than BIG versions of each because postgres returns BIG ones as strings

+++ migrations/sqls/20210309162402-orders-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "orders" (
-    id BIGSERIAL NOT NULL PRIMARY KEY,
-    user_id BIGINT REFERENCES users(id),
+    id SERIAL NOT NULL PRIMARY KEY,
+    user_id INT REFERENCES users(id),
     order_complete BOOLEAN
 );
